requirements.txt
fastapi==0.110.1
pandas==1.5.3
pydantic[email]==1.10.14
email-validator==1.1.3
PyMySQL==1.1.0
SQLAlchemy==2.0.28
uvicorn==0.29.0
Werkzeug==3.0.2
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
requests==2.31.0
openpyxl==3.1.2
tqdm==4.66.2
aiomysql==0.2.0
python-dotenv==1.0.1